src/data_ingestion/requirements.txt
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
pypdf
